+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public abstract class Command {
     public abstract String execute(TaskList tasks, Ui ui, Storage storage);
