+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class MarkCommand extends Command {
     private final int targetIndex;
 
@@ -8,7 +15,7 @@ public class MarkCommand extends Command {
     public String execute(TaskList tasks, Ui ui, Storage storage) {
         Task t = tasks.markTask(targetIndex);
         storage.save(tasks);
-        return "Great job! I've marked the task as completed:\n" + ui.tab(t.toString());
+        return "Great job! I've marked the duke.task as completed:\n" + ui.tab(t.toString());
     }
 
     public boolean isExit() {
