+++ src/main/java/duke/commands/UnmarkCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class UnmarkCommand extends Command {
     private final int targetIndex;
 
@@ -9,7 +16,7 @@ public class UnmarkCommand extends Command {
     public String execute(TaskList tasks, Ui ui, Storage storage) {
         Task t = tasks.unmarkTask(targetIndex);
         storage.save(tasks);
-        return "No problem, I've marked the task as uncompleted:\n" + ui.tab(t.toString());
+        return "No problem, I've marked the duke.task as uncompleted:\n" + ui.tab(t.toString());
     }
 
     public boolean isExit() {
