+++ src/main/java/duke/exceptions/InvalidDateTimeException.java
@@ -1,4 +1,6 @@
-public class InvalidDateTimeException extends DukeException{
+package duke.exceptions;
+
+public class InvalidDateTimeException extends DukeException {
     public InvalidDateTimeException() {
         super("Please enter the date and/or time in the specified format:\n" +
                 "yyyy-MM-dd HHmm\n" +
