+++ src/main/java/duke/task/TaskList.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.util.ArrayList;
 
 public class TaskList {
@@ -43,7 +45,7 @@ public class TaskList {
             s.append("    ");
             s.append(i + 1).append(". ");
             s.append(tasks.get(i));
-            // no new line after last task
+            // no new line after last duke.task
             if (i + 1 < tasks.size()) {
                 s.append("\n");
             }
