+++ text-ui-test/EXPECTED.TXT
@@ -17,31 +17,31 @@
     ============================================================
 
     ============================================================
-    Task added:
+    duke.task.Task added:
     [T][ ] read book
     ============================================================
     ============================================================
-    Task added:
+    duke.task.Task added:
     [D][ ] do homework (by: Saturday 3pm)
     ============================================================
     ============================================================
-    Task added:
+    duke.task.Task added:
     [E][ ] zoom meeting (at: Monday 2-3pm)
     ============================================================
     ============================================================
-    Great job! I've marked the task as completed:
+    Great job! I've marked the duke.task as completed:
     [T][X] read book
     ============================================================
     ============================================================
-    Great job! I've marked the task as completed:
+    Great job! I've marked the duke.task as completed:
     [D][X] do homework (by: Saturday 3pm)
     ============================================================
     ============================================================
-    Great job! I've marked the task as completed:
+    Great job! I've marked the duke.task as completed:
     [E][X] zoom meeting (at: Monday 2-3pm)
     ============================================================
     ============================================================
-    No problem, I've marked the task as uncompleted:
+    No problem, I've marked the duke.task as uncompleted:
     [D][ ] do homework (by: Saturday 3pm)
     ============================================================
     ============================================================
